+++ django_typer/templates/shell_complete/bash.sh
@@ -33,7 +33,7 @@
     response=()
     while IFS= read -r line; do
         response+=("$line")
-    done < <( $1 {{ django_command }} --shell bash ${settings_option:+${settings_option}} ${pythonpath_option:+${pythonpath_option}} {{ color }} complete {{ fallback }} "${COMP_WORDS[*]}" )
+    done < <( $1 {{ django_command }} --shell bash ${settings_option:+${settings_option}} ${pythonpath_option:+${pythonpath_option}} {{ color }} complete {{ fallback }} "${COMP_WORDS[*]}" "$COMP_POINT" )
     
     COMPREPLY=()
     set_mode=true