+++ django_typer/types.py
@@ -110,9 +110,9 @@ Settings = Annotated[
     Option(
         help=cast(
             str,
-            _(
+            (
                 "The Python path to a settings module, e.g. "
-                '"myproject.settings.main". If this isn\'t provided, the '
+                '"myproject.settings.main". If this is not provided the '
                 "DJANGO_SETTINGS_MODULE environment variable will be used."
             ),
         ),